+++ samples/Samples.ContextDemo/Event.cs
@@ -1,0 +1,11 @@
+namespace Machinery;
+
+internal enum Event
+{
+    None = 0,
+
+    // The same event used for both opening and closing.
+    Interact,
+    Lock,
+    Unlock
+}